move cards to campaign view, use grid column not area
