+++ client/visual/grid.js
@@ -4,7 +4,7 @@ import shortId from '@quarterto/short-id';
 const bleed = `__bleed_${shortId()}`;
 
 export const Bleed = styled.div.attrs({className: bleed})`
-	grid-area: bleed;
+	grid-column: bleed;
 `;
 
 export default styled.div`
@@ -16,6 +16,6 @@ export default styled.div`
 	grid-gap: 1em;
 
 	> :not(.${bleed}) {
-		grid-area: content;
+		grid-column: content;
 	}
 `;
